pridané hodnotenia použivateľa

+++ public/js/user.js
@@ -1,29 +1,70 @@
-$(document).ready(function() {
+function loadItems(owned) {
     $.ajax({
         type: "GET",
-        url: "?c=user&a=getUserItems",
+        url: "?c=user&a=getUserItems&p=" + $("#user-id").text(),
         dataType: "json",
         success: function(response) {
             if (response.length == 0) {
-                $("#ads").append("<p>Nemáte zverejnené žiadne inzeráty</p>");
+                owned ? $("#ads").append("<p>Nemáte zverejnené žiadne inzeráty</p>") : $("#ads").append("<p>Užívateľ nemá zverejnené žiadne inzeráty</p>");
             } else {
                 let html = "";
                 
                 for (let i = 0; i < response.length; i++) {
                     html += "<div class='item'>"
-                         + "<p>" + response[i].title + "</p>"
-                         + "<div class='icons'>"
-                         + "<a href='?c=editor&p=" + response[i].id + "'><i class='fas fa-edit'></i></a>" 
-                         + "<a href='?c=editor&a=delete&p=" + response[i].id + "'><i class='fas fa-trash'></i></a>"
-                         + "</div>"; 
-                         + "</div>";
+                         + "<p>" + response[i].title + "</p>";
+
+                    if (owned) {
+                        html += "<div class='icons'>"
+                              + "<a href='?c=editor&p=" + response[i].id + "'><i class='fas fa-edit'></i></a>" 
+                              + "<a href='?c=editor&a=delete&p=" + response[i].id + "'><i class='fas fa-trash'></i></a>"
+                              + "</div>"
+                              + "</div>";
+                    } else {
+                        html += "</div>";
+                    }
                 }
 
                 $("#ads").append(html);
             }
         }
     });
-})
+}
+
+function loadReviews() {
+    $.ajax({
+        type: "GET",
+        url: "?c=user&a=getUserReviews&p=" + $("#user-id").text(),
+        dataType: "json",
+        success: function(response) {
+            if (response.length == 0) {
+                $("#reviews").append("<p>Žiadne hodnotenia</p>");
+            } else {
+                for (let i = 0; i < response.length; i++) {
+                    let review = response[i];
+                    let id = "";
+
+                    $.ajax({
+                        type: "GET",
+                        url: "?c=user&a=getUserDetails&p=" + review.author,
+                        dataType: "json",
+                        async: false,
+                        success: function(tmp_response) {
+                            id = review.author;
+                            review.author = tmp_response.name;
+                        }
+                    });
+
+                    let html = "<div class='review'>"
+                             + "<a href='?c=user&a=profile&p=" + id + "'>" + review.author + "</a>"
+                             + "<p>" + review.content + "</p>"
+                             + "</div>";
+
+                    $("#reviews").append(html);
+                }
+            }
+        }
+    });
+}
 
 function updatePhone() {
     let numb = $("#user-phone");
@@ -53,4 +94,9 @@ function updatePhone() {
         input.val($("#user-phone").text());
         input.show();
     }
+}
+
+function showEditor() {
+    $(".review").hide();
+    $("#editor").show();
 }
